Revert "Merge pull request #59 from shiftcommerce/hotfix/3826_sellable_flag_lag"
This reverts commit 84f4ac40e3a10d709a18075a0c9cc6a792ad034d, reversing
changes made to 3f153a29ff5e4e5ea714b1bf9d9b40101e6296c1.

The changes originally made by this branch do not appear to work.  The test suite was failing in the platform (10 failures) and the fix was not too obvious, I believe it is
the way that SHQL is being merged - you cannot just merge an SHQL hash with another - it wont know what to do with the 2 sides - whether to "and" or "or" them.